requirements.txt
pyserial~=3.5
peewee~=3.14.3
requests~=2.25.1
gpiozero~=1.5.1
Pillow~=8.1.2
xmltodict~=0.12.0
tflite-runtime